исправления

+++ src/main/java/ru/kaInc/shelterbot/service/implementation/PhotoServiceImpl.java
@@ -1,9 +1,12 @@
 package ru.kaInc.shelterbot.service.implementation;
 
+import jakarta.persistence.EntityExistsException;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartException;
 import org.springframework.web.multipart.MultipartFile;
 import ru.kaInc.shelterbot.model.Photo;
 import ru.kaInc.shelterbot.repo.PhotoRepo;
@@ -30,14 +33,29 @@ public class PhotoServiceImpl implements PhotoService {
 
     @Override
     public Optional<Photo> findPhotoById(Long id) {
+        if (photoRepo.findById(id) == null) {
+            throw new EntityNotFoundException(String.format("Photo with id %s not found", id));
+        }
         return photoRepo.findById(id);
     }
 
     @Override
+    @Transactional
     public Long upLoadPhoto(Long id, MultipartFile multipartFile) throws IOException {
+        if (multipartFile == null) {
+            throw new MultipartException("Photo was null");
+        }
+        if (multipartFile.getSize() > 1024 * 600) {
+            throw new OutOfMemoryError("Photo size is too big");
+        }
+        if (photoRepo.findById(id).isPresent()) {
+            throw new EntityExistsException(String.format("Photo with id %s exists", id));
+        }
         Photo newPhoto = new Photo();
 
         newPhoto.setId(id);
+        newPhoto.setMediaType(multipartFile.getContentType());
+        newPhoto.setFileSize(multipartFile.getSize());
         newPhoto.setData(generateDataForDB(multipartFile));
 
         photoRepo.save(newPhoto);
@@ -61,6 +79,9 @@ public class PhotoServiceImpl implements PhotoService {
     @Override
     public void deletePhoto(Long photoId) {
         logger.info("Photo deleted");
+        if (photoRepo.findById(photoId).isEmpty()) {
+            throw new EntityNotFoundException(String.format("Photo with id %s not found", photoId));
+        }
         photoRepo.deleteById(photoId);
     }
